Support for arbitrary data types
Not efficient[1], but that mostly needs work on the arbitrary crate, as
opposed to work on libfuzzer-sys.

[1]: fuzzing struggles with e.g.

    fuzz_target!(|data: String| { if data == "banana" { panic!() } });

and takes quite a while to find an input that fails the check (compared
to fuzzing against raw bytes). It seems to me that to take full
advantage of fuzzing structured data, one would have to adjust libFuzzer
somehow or, perhaps, rewrite libFuzzer in rust having precisely
structured fuzzing in mind.