CI: build with a single codegen unit
This is not ideal, but it does provide two things:

1. It fixes bizarre linker errors about missing `sancov` symbols.

2. It allows LLVM to do inlining that it otherwise refuses to do. For some
reason, when sanitizers are enabled, LLVM refuses to inline across codegen
units. This is a problem because trivial methods like `Vec::len` won't be
inlined, resulting in 100x slowdowns.

`cargo fuzz` already restricts its builds to a single codegen unit, so we might
as well do the same thing in CI here.