Add changes suggested from code review.
Docs: make it explicit that we're ignoring the return value of the
function under test. Add comments from libfuzzer explaining why one
might want to keep inputs out of the corpus.

Convert From<Corpus> to i32 to a pub fn to_libfuzzer_code() that is
impl on Corpus to avoid accidental conversion.